+++ packages/react-filerobot-image-editor/src/components/Topbar/RedoButton.jsx
@@ -8,7 +8,7 @@ import { REDO } from 'actions';
 import { useStore } from 'hooks';
 import { StyledHistoryButton } from './Topbar.styled';
 
-const RedoButton = ({ margin, buttonIconPrimaryColor }) => {
+const RedoButton = ({ margin }) => {
   const { dispatch, hasRedo = false, t } = useStore();
   const dispatchRedo = useCallback(() => {
     dispatch({ type: REDO });
@@ -21,7 +21,6 @@ const RedoButton = ({ margin, buttonIconPrimaryColor }) => {
       size="sm"
       onClick={hasRedo ? dispatchRedo : undefined}
       disabled={!hasRedo}
-      buttonIconPrimaryColor={buttonIconPrimaryColor}
       title={t('redoTitle')}
       margin={margin}
     >
@@ -32,12 +31,10 @@ const RedoButton = ({ margin, buttonIconPrimaryColor }) => {
 
 RedoButton.defaultProps = {
   margin: undefined,
-  buttonIconPrimaryColor: false,
 };
 
 RedoButton.propTypes = {
   margin: PropTypes.string,
-  buttonIconPrimaryColor: PropTypes.bool,
 };
 
 export default RedoButton;